Progress : 1. Analyzed reason for kprobe not registering. Turns out, we had disabled a check previously that originally doesnt allows installing kprobes on non-core kernel text. Following the observation, those checks are disabled again in current kernel version until we think about it later. 2. Kprobes are now registering after succesfully indentifying call insns. 3. Next steps is to identify the call address (which is saved somewhere in the struct kprobe) and make the pre-handler use kallsyms table to find out which helper it was pointing to.   3.1 Then depeding on return type, return an acceptable failure value. (0,-1, etc.)
Next commits will be about a testing infra where a syscall to verify a bpf program would involve
calling a  function which will one-by-one patch call insns and pass them to the
verifier for analysis. This way, all possible termination scenarios will be convered to guarantee
that no properties will be violated when abrupt termination is requested.