Enable per fentry program stack offset
Add a field to bpf_link struct to store the stack offset.
Change trampoline to add space for this offset.
Modify helper functions to access the new spot.

The flow will then be get the offset in the helper, then resolve that to
the actual stack storage location.

This is needed because we may have many pairwise or not programs
attached to a single hook. This allows each to have their own stack
storage if they need. Also ensures the helper accesses the right one.